* Articles may offer questions and answers to test one's knowledge.
article.html fetches its CSS (master.css) and JavaScript (master.js) code from the
server instead of having these inlined. The reason is mainly practical. This is
possible because the relevant paths are met with website#static(name).

master.js introduces a first step towards the notion of actors. The entire article is
now viewed as an actor. This behaviour includes the following: if there are exercises
in the article, then a Q&A button is displayed in the top-right corner.

If the user clicks on this button, it is equivalent to Article#review(article).
As a consequence, the content of the article is hidden and a review window becomes
visible. Exercises may now be reviewed in order, as flashcards would be.

Appropriate definitions and algorithms allow the exercises to be stored in the HTML
file as exported from the source org file and then re-used for the dynamic behaviour
described above.

Schematically, starting just after the browser received an article from the website:
- GET /static/master.css
- GET /static/master.js
  - Article#mk()
    - content :≡ Content#mk()
    - Content#show(content)
    - exercises : List(Exercise) ≡ list-exercise()
    - empty?(exercises)
      - #t → ■
      - #f →
        - review :≡ Review#mk(exercises)
        - Review#hide(review)
        - qa-btn :≡ QABtn#mk()
        - QABtn#show(qa-btn)
        - Actor#subscribe(qa-btn self)
          - After qa-btn is clicked then the content is hidden and the review actor
            is (re)initialized and shown.
        - Actor#subscribe(review self)
          - After review has finished or has been stopped, the review actor is hidden
            and the content shown.